Project_3_RealTime_Streaming/requirements.txt
kafka-python>=2.0.2
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=0.19.0
pytest>=7.0.0
